[BUGFIX] Initialize page generator in dummy page generation script
Without this, e.g. $TSFE->linkVars is not initialized in
USER_INT (uncached extbase actions) if the page cache is completely
empty due to the missing call to PageGenerator::pagegenInit().

I copied the whole block from index_ts.php including the include
lines to have a consistent environment in that situation.