[BUGFIX] Allow rendering of pages context on a root page
If rendering a path without a record context, the renderer
automatically resolves the current page as context.

However if this page is a root page (pid is 0) then
the RECORD content object refuses to render this page
if `dontCheckPid` is not set.

Catch this situation and set this configuration if needed.