[FOLLOWUP] Allow rendering of pages context on a root page
As @ffrewer correctly pointed out, the check for the root page was bogus.
We need to evaluate the id of the record, which we're going to render
and check if this page is the root page in the current root line.

We ignore the case to render a root page of another root line (multi domain setup)