[!!!][BUGFIX] Respect different rendering for content elements
If a plugin is registered as content element, the TypoScript rendering path
is different than when being registered as regular plugin.

We now respect that and additionally throw an exception in case the
rendering path cannot be resolved.

This can be considered as breaking. Before the URL returned nothing
and the page generating this URL kept working. Now the page generating
the invalid URL will be broken as an exception is thrown in this case.

Fixes #19