Rename AjaxActionViewHelper to ActionViewHelper
To be consistent with the link view helper and also
to clarify, that the URI generated might be used for
different purposes than just Ajax, like download links for example,
we rename the view helper but keep the old one to provide
backwards compatibility.

Usage of the old one is deprecated though.