Adds value for missing argument "extensionName"
In TyposcriptRenderingUri.php, the argument "extensionName" is passed to
$uriBuilder->uriFor (line 109). When using the t:link.cObject view helper,
this results in an exception:
Argument 1 passed to TYPO3\CMS\Extbase\Service\ExtensionService::getTargetPidByPlugin()
must be of the type string, null given.

Since the plugin name isn't needed for the cObject view helper,
we set the extension name to a dummy value.